Handle black and white pawn moves
Prompt: Please extend the code to accept lowercase to indicate a black piece and uppercase
for white. Handle moves for black pawns ('p') and white pawns ('P'). Add the two-square
initial move as well, but not capturing or en passant. As the function is getting big,
break the moves for each of the piece types out in its own separate function.  # Please
enter the commit message for your changes. Lines starting # with '#' will be ignored,
and an empty message aborts the commit.