Compute possibleCaptures as well as possibleMoves
Prompt: Write a function `possibleCaptures` similar to `possibleMoves`, but using the
appropriate rules for pawns to capture pieces, still ignoring the en passant rule. Reuse
the existing functions for piece moves where possible.

Prompt: The code for `possibleCaptures` is incorrect: it invokes the piece move functions
with a `captures` argument, but these functions return a set of squares instead of modifying
a passed in reference to a set of squares. Please change `possibleCaptures` to directly
return the result of the move functions.

Prompt: Please include the definition of `queenMoves`, as this function was not yet
known. There are some functions that also could benefit from the `queenMoves` function. Give
their updated definitions as well.

Prompt: For the `possibleMoves` function, still avoid using `std::tolower` by adding extra
cases to the switch statement.