Add availableCaptures
Write a function `availableCaptures` that takes a `ChessBoard` and a active color ('w'
or 'b') , and returns a  set of moves that includes the possible captures of each of our
pieces, excluding self capture  (we already have a piece on the `to` square) or moves that
would move through pieces. We still exclude en passant capture, and do not exclude moves
that might check our king. This should be analogous to `availableMoves`, but for captures.