Add function to determine if a piece is attacked
Prompt: Define a function `isAttacked`, that given a `ChessBoard` and a `Square` returns
a bool indicating whether the square has a piece that is attacked by any of the opponent's
pieces. We should be able to use the already defined `availableCaptures` function.

Prompt: Rather than passing in the `activeColor`, derive it from the piece at the given
square. If the square is empty, it is not attacked.