Redo opening book handling
Now convert the PGN files into a CSV that is
smaller and faster to process. Improve move
selection by adding Dirichlet priors to the move
probabilities, and allow setting a temperature for
more variety in book move selection. Additionally
use a bonus for common opening moves to appear
more human-like.