Correctly determine player to move for time control
There was a major bug in time control, where it would only consider the start position for the player's turn, not the additional moves playes. For UCI games, the position is almost always the startpos, with all other moves passed as moves argument. Effectively this was always assuming the engine would be playing as white leading to faulty time control.